backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=42.0.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
alembic>=1.14.0
email-validator>=2.2.0
metaapi-cloud-sdk>=27.0.0
